+++ ltexls-core/src/main/java/org/bsplines/ltexls/server/LtexWorkspaceService.java
@@ -8,10 +8,15 @@
 package org.bsplines.ltexls.server;
 
 import com.google.gson.JsonObject;
+import com.sun.management.OperatingSystemMXBean;
+import java.lang.management.ManagementFactory;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.time.Duration;
+import java.time.Instant;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
@@ -27,9 +32,10 @@ import org.eclipse.lsp4j.WorkspaceSymbolParams;
 import org.eclipse.lsp4j.services.WorkspaceService;
 
 class LtexWorkspaceService implements WorkspaceService {
-  @NotOnlyInitialized LtexLanguageServer languageServer;
-
   private static final String checkDocumentCommandName = "ltex.checkDocument";
+  private static final String getServerStatusCommandName = "ltex.getServerStatus";
+
+  @NotOnlyInitialized LtexLanguageServer languageServer;
 
   public LtexWorkspaceService(@UnknownInitialization LtexLanguageServer languageServer) {
     this.languageServer = languageServer;
@@ -54,6 +60,8 @@ class LtexWorkspaceService implements WorkspaceService {
   public CompletableFuture<Object> executeCommand(ExecuteCommandParams params) {
     if (params.getCommand().equals(checkDocumentCommandName)) {
       return executeCheckDocumentCommand((JsonObject)params.getArguments().get(0));
+    } else if (params.getCommand().equals(getServerStatusCommandName)) {
+      return executeGetServerStatusCommand((JsonObject)params.getArguments().get(0));
     } else {
       return CompletableFuture.completedFuture(false);
     }
@@ -108,6 +116,42 @@ class LtexWorkspaceService implements WorkspaceService {
     });
   }
 
+  public CompletableFuture<Object> executeGetServerStatusCommand(JsonObject arguments) {
+    if (this.languageServer == null) {
+      return failCommand(Tools.i18n("languageServerNotInitialized"));
+    }
+
+    final long processId = ProcessHandle.current().pid();
+    final double wallClockDuration = Duration.between(
+        this.languageServer.getStartupInstant(), Instant.now()).toMillis() / 1000.0;
+    @Nullable Double cpuDuration = null;
+    @Nullable Double cpuUsage = null;
+    final double totalMemory = Runtime.getRuntime().totalMemory();
+    final double usedMemory = totalMemory - Runtime.getRuntime().freeMemory();
+
+    try {
+      OperatingSystemMXBean operatingSystemMxBean =
+          (OperatingSystemMXBean)ManagementFactory.getOperatingSystemMXBean();
+      cpuUsage = operatingSystemMxBean.getProcessCpuLoad();
+      if (cpuUsage == -1) cpuUsage = null;
+      long cpuDurationLong = operatingSystemMxBean.getProcessCpuTime();
+      cpuDuration = ((cpuDurationLong != -1) ? (cpuDurationLong / 1e9) : null);
+    } catch (ClassCastException e) {
+      // do nothing
+    }
+
+    JsonObject jsonObject = new JsonObject();
+    jsonObject.addProperty("success", true);
+    jsonObject.addProperty("processId", processId);
+    jsonObject.addProperty("wallClockDuration", wallClockDuration);
+    if (cpuUsage != null) jsonObject.addProperty("cpuUsage", cpuUsage);
+    if (cpuDuration != null) jsonObject.addProperty("cpuDuration", cpuDuration);
+    jsonObject.addProperty("usedMemory", usedMemory);
+    jsonObject.addProperty("totalMemory", totalMemory);
+
+    return CompletableFuture.completedFuture(jsonObject);
+  }
+
   private static CompletableFuture<Object> failCommand(String errorMessage) {
     JsonObject jsonObject = new JsonObject();
     jsonObject.addProperty("success", false);
@@ -116,6 +160,6 @@ class LtexWorkspaceService implements WorkspaceService {
   }
 
   public static List<String> getCommandNames() {
-    return Collections.singletonList(checkDocumentCommandName);
+    return Arrays.asList(checkDocumentCommandName, getServerStatusCommandName);
   }
 }